refactor: clients are now called agents
