+++ sorrydb/cli/run_llm_agent.py
@@ -4,30 +4,39 @@ import argparse
 import json
 import logging
 import sys
-from pathlib import Path
 
-from sorrydb.clients.rfl_client.rfl_client import process_sorries_json
+from sorrydb.agents.llm_agent.llm_agent import LLMAgent
 
 
 def main():
-    parser = argparse.ArgumentParser(description="Reproduce a sorry with REPL.")
+    parser = argparse.ArgumentParser(description="Solve sorries using LLMClient")
+
     parser.add_argument(
-        "--sorry-file",
+        "--sorry-db",
         type=str,
-        required=True,
-        help="Path to the sorry JSON file",
+        default="https://raw.githubusercontent.com/austinletson/sorry-db-data/refs/heads/master/sorry_database.json",
+        help="URL to the sorry database JSON file",
     )
+
     parser.add_argument(
-        "--output-file",
+        "--out",
         type=str,
-        required=True,
-        help="Path to the output JSON file",
+        default="llm_proofs.json",
+        help="Path to the output JSON file (default: llm_proofs.json)",
     )
+
     parser.add_argument(
-        "--lean-data",
+        "--model-json",
         type=str,
         default=None,
-        help="Directory to store Lean data (default: use temporary directory)",
+        help="Path to the model config JSON file (default: None)",
+    )
+
+    parser.add_argument(
+        "--lean-dir",
+        type=str,
+        default="lean_data",
+        help="Directory to store Lean data (default: lean_data)",
     )
 
     parser.add_argument(
@@ -37,6 +46,7 @@ def main():
         choices=["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"],
         help="Set the logging level (default: INFO)",
     )
+
     parser.add_argument(
         "--log-file", type=str, help="Log file path (default: output to stdout)"
     )
@@ -47,24 +57,21 @@ def main():
     log_kwargs = {
         "level": getattr(logging, args.log_level),
         "format": "%(asctime)s - %(name)s - %(levelname)s - %(message)s",
+        "datefmt": "%Y-%m-%d %H:%M:%S",  # No ms
     }
+
     if args.log_file:
         log_kwargs["filename"] = args.log_file
-    logging.basicConfig(**log_kwargs)
 
+    logging.basicConfig(**log_kwargs)
     logger = logging.getLogger(__name__)
 
-    # Convert file names arguments to Path
-    sorry_file = Path(args.sorry_file)
-    output_file = Path(args.output_file)
-    lean_data = Path(args.lean_data) if args.lean_data else None
-
-    # Process the sorry JSON file
+    # Process the sorry DB using the LLMClient
     try:
-        logger.info(f"Processing sorry file: {sorry_file}")
-        process_sorries_json(sorry_file, output_file, lean_data)
+        logger.info(f"Solving sorry db at {args.sorry_db} using LLMAgent.")
+        agent = LLMAgent(args.model_json, args.lean_dir)
+        agent.solve_sorry_db(args.sorry_db, args.out)
         return 0
-
     except FileNotFoundError as e:
         logger.error(f"File not found: {e}")
         return 1
